video-production-platform/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
chromadb==0.4.22
httpx==0.25.2
python-multipart==0.0.6
edge-tts==6.1.9
mutagen==1.47.0
hypothesis==6.92.2
pytest==7.4.3
pytest-asyncio==0.23.2
aiosqlite==0.19.0
jinja2==3.1.2
